app/requirements.txt
tensorflow==2.21.0
numpy
opencv-python-headless
Pillow
